requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
channels>=4.0,<5.0
opencv-python>=4.8
mediapipe==0.10.14
numpy>=1.26
faster-whisper>=1.0.3
